Use unique callbacks to avoid JSONP race condition
Unique callback names are only really required when loading multiple HARP
files in a single browser window. I.e. they are not required for the APIs
that result in separate IFRAMEs because the IFRAMEs have their own window
scope. The reason is the jQuery JSONP handling uses a global callback
function attached to the `window` object, and so multiple same-named
callbacks on the same `window` object will clash.

Several `*.html.php` test fixtures have been updated to use a specific
unique callback name. This change required updates to the test HARP files
- giving them unique callback names.

`har.js` now looks for a `data-callback` attribute, which is the callback
name to use for the associated HARP file.