Put Stylesheets at the Top
https://developer.yahoo.com/performance/rules.html#css_top

I noticed a couple of IE9 tests sporadically failing, and when I looked at
what was showing in the IE9 browser, I could see an unstyled harviewer
page (CSS seemed to have failed to load).

This change is both a performance best practice, and also seems to stop
this odd and occasional IE9 behaviour.