Upgrade RequireJS to 2.1.18 (and r.js/text/i18n)
- RequireJS 2.1.18
- r.js 2.1.18
- text 2.0.14
- i18n 2.0.2

The module definitions have been changed as in the following example:

From
  require.def("harPreview", [

to
  define("harPreview", [

This is not ideal as define() should be used without an explicit module
name:

- http://requirejs.org/docs/api.html#modulename

The Ant build.xml file now downloads r.js on-demand from GitHub. "r.js"
is the new version of the RequireJS build tool.

- https://github.com/jrburke/r.js

This change means the old requirejs folder can probably be deleted from
the repo.

The build does not query the O/S any more.  The new way to build uses an
Ant property called `rjs.runner`.  The default value for this is 'node',
and executing `ant` from the command line will use Node.js to run the
"r.js" build.  Of course this requires Node.js being installed.

If Java is preferred, use `ant -Drjs.runner=java` from the command line.

The speed of the build may vary significantly depending on choice of
Node.js/Java and optimiser (uglify, uglify2, closure, etc).