Start to add code coverage reports
- Add Lcov and LcovHtml reporters to intern config.
- Add new 'tests/unit/core/lib' and 'tests/unit/core/cookies' unit tests.

/lcov.info file and /html-report/ folder will be generated.

For the code coverage to work with functional tests, the proxyUrl in the
Intern config needs to be a URL that is reachable from the Selenium test
nodes (i.e. not localhost if the Selenium nodes are on a different host to
the Intern proxy).

Both the $harviewer_base and $test_base URLs in config.php, and the
harviewer.harViewerBase and harviewer.testBase URLs in the Intern config
need to point to the proxyUrl to get coverage information.

E.g. (in /selenium/tests/config.php):

$harviewer_base = "http://192.168.1.80:9000/webapp/";
$test_base = "http://192.168.1.80:9000/selenium/";

and (in /tests/intern.js):

    harviewer: {
        harViewerBase: 'http://192.168.1.80:9000/webapp/',
        testBase: 'http://192.168.1.80:9000/selenium/'
    },

Config value `excludeInstrumentation` has also been improved to exclude
all code except the application code.