Add Image and ExternalImage tabs
ImageTab tries to use the response.content.text and
response.content.encoding fields to create a data URI for the src
attribute of an IMG tag in order to display an image preview. This means
no HTTP request is sent for the image.

ExternalImageTab uses request.url for the src attribute of an IMG tag in
order to display an image preview.  A HTTP request will be sent.

Adds appDriver.js for common HAR Viewer-specific WebDriver actions.

Fixes #23