Update SyntaxHighlighter from 1.5.1 to 3.0.83
Fixes an issue with some syntax highlighting hanging the browser (e.g.
trying to syntax highlight the minified version of require.js).

Version 3.0.83 located at:
- https://github.com/alexgorbatchev/syntaxhighlighter/tree/3.0.83/

Replaced SyntaxHighlighter.css with contents from:
- https://github.com/alexgorbatchev/syntaxhighlighter/blob/3.0.83/styles/shCore.css
- https://github.com/alexgorbatchev/syntaxhighlighter/blob/3.0.83/styles/shCoreDefault.css

Replaced shCore.js with contents from:
- https://github.com/alexgorbatchev/syntaxhighlighter/blob/3.0.83/scripts/XRegExp.js
- https://github.com/alexgorbatchev/syntaxhighlighter/blob/3.0.83/scripts/shCore.js
- https://github.com/alexgorbatchev/syntaxhighlighter/blob/3.0.83/scripts/shBrushJScript.js
- https://github.com/alexgorbatchev/syntaxhighlighter/blob/3.0.83/scripts/shBrushCss.js
- https://github.com/alexgorbatchev/syntaxhighlighter/blob/3.0.83/scripts/shBrushXml.js

To make this work, the CSS class of the element containing the code to
highlight changes from something like:

  class="javascript:nocontrols:nogutter:"

to:

  class="toolbar: false; brush: javascript"

The API call to do the highlighting changes from:

  dp.SyntaxHighlighter.HighlightAll(pre)

to:

  dp.SyntaxHighlighter.highlight(pre)

The SchemaTab uses SyntaxHighlighter, so that also had to change.

Fixes #81.