Add tree viewer for JSON and XML
Updates xml.har to have response content XML that has attributes and
comments in order to test all features of the XML tree viewer.

Adds rss.har that gives an example of "application/rss+xml" content so we
can make sure this is treated like XML.

Updates json.har to have response content JSON that is more complex in
order to test all features of the JSON tree viewer.

Fixes #62.