Add serverIPAddress and connection requestList columns
As there are now more columns, if they are all visible then the chances of
being able to read all their values is reduced.  Therefore, add HTML title
attributes to all columns (that don't already have a specialised tooltip)
so that a user can hover over the column to see the value.

Fixes #60.
Refs #42.