Update docs/About tab
Make example URLs relative to the current URL.

Also add examples/inline-scripts-block.harp.js file which is a copy of
inline-scripts-block.harp, but with a .js extension.

This is so we can load this HARP file from GitHub pages and have it served
with a JavaScript content type (rather than text/plain as served from
GitHub), and therefore we can use this HARP in our cross-domain examples
in the About tab.