Add URL parameter aliases. har=path, harp=inputUrl
Refactor harModelLoader to be able to load a list of HARs (via XHR) and a
list of HARPs (via JSONP).

Cross-domain HAR loading is possible with CORS and cross-domain XHR, so
JSONP is not required in all cross-domain cases.

Use the "inputUrl" or "harp" URL parameters to load HARP (JSONP) files
(same-domain or cross-domain, it doesn't matter).

Use the "path" or "har" URL parameters to load HAR files (same-domain or
cross-domain, but cross-domain loading requires CORS).

The "har" parameter is superior to "path" because it will load *all* the
hars specified by "har" parameters.  "path" was limited to loading just
the first "path" parameter.

Also, the "path" parameter was inconsistent in that it would sometimes
expect a file to be a HAR and other times a HARP (depending on presence of
"baseUrl" parameter").

The "har" and "harp" parameter names are more obvious. Want to load a HAR,
use "har". Want to load a HARP, use "harp". Want to load a mixture of HARs
and HARPs, use both parameters.