Fix object.js bind() method
The bind() method had this JSDoc:

// fn, thisObject, args => thisObject.fn(args, arguments);

But the bound function was not being invoked with the 'bound' arguments
before the 'runtime' arguments.

Fix the order of the arguments, and change the code that relied on the
opposite order.

The new order is the same as that used by native bind():

https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_objects/Function/bind

Unit test added.