Fix #98
1) Use the 'new-style' HAR and HARP loading for previews.

Old style:

An element's "data-har" attribute is inspected to see if it's
an absolute URL or not.  If it's absolute it's assumed to be a HARP
file (cross-domain).  If it's not absolute, it's assumed to be a HAR
file (same-domain).

New-style:

If an element has a "data-har-url" attribute, this is treated as a HAR
file whether or not it is same-domain.  The user must rely on correct
CORS settings to be able to read cross-domain HAR files.

If an element has a "data-harp-url" attribute, this is treated as a HARP
file whether or not it is same-domain.

2) Replace preview.php with preview.html.