Update text of issue link
Hopefully a new version can be released before too long so that the main website is updated. Right now http://www.softwareishard.com/har/viewer/ is still linking out to code.google.com for the issues, and that's the page that came up from a Google search. Had to do some digging to find the Github page.