Rewrite functional tests for Intern 4 ...
1) We now have a separate config module to store the functional test
config. We run into problems require-ing the old intern.js module as it
uses dojo/has which wasn't loading properly. As we only need a couple of
config properties it is easier to separate these out.

2) Any const declarations must live within the define() functions to avoid
polluting the global scope and clashing with each other (e.g.
`registerSuite` and `assert`).

3) pollUtil is a default export, and so if we require it normally we get an
object with a default property returned, e.g.

{ default: function pollUntil() ... }

So we require leadfoot itself and pull the value from that.

The same applies with the `keys` function.

4) teardown lifecycle method is replaced by afterEach

5) clearCookies() is not working in edge driver 17134, so
"testValidateCheckbox" test fails with this driver and Edge 42.17134.1.0.

See
https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/14838528/

6) Add various configs to intern.json so test with various browsers.