Better support for cross-domain tests
Using switchToFrame means we can use selectors in the context of a
child frame, which means the frame need not be the same domain as the
top-level window's domain.

Before, the code relied on the top-level window and the frame being the
same domain, so that JavaScript executed in the context of the top-level
window would work within the frame, without cross-domain issues.

Changes to:

- testEmbeddedPreview
- testIssue39
- testEmbeddedInvalidPreview
- DriverUtils.querySelectAllInFrameAndReturnLengthOrNull no longer needed