add example of embedding. see #130
A link to the new embed.html page is given on the home page and the
about page.

Fix styling issue where the preview mode does not display tabs like
"Response" and "JSON" correctly because of missing styles.

Update harViewer.css and harPreview.css to try and have the same CSS
imports on the same lines, for ease of comparison between the two.